+++ Scroller/Scroller.js
@@ -31,6 +31,7 @@ import Scrollbar from '../useScroll/Scrollbar';
 import ScrollbarPlaceHolder from '../useScroll/ScrollbarPlaceHolder';
 import Skinnable from '../Skinnable';
 
+import {EditableShape, EditableWrapper} from './EditableWrapper';
 import useThemeScroller from './useThemeScroller';
 
 const nop = () => {};
@@ -77,9 +78,12 @@ let Scroller = ({'aria-label': ariaLabel, hoverToScroll, ...rest}) => {
 	} = scrollContentWrapperProps;
 
 	const {
+		editableWrapperProps,
 		focusableBodyProps,
 		themeScrollContentProps
 	} = useThemeScroller(rest, {...scrollContentProps, className: classnames(className, scrollContentProps.className)}, id, isHorizontalScrollbarVisible, isVerticalScrollbarVisible);
+
+	const {children, direction, editable} = rest;
 
 	// To apply spotlight navigableFilter, SpottableDiv should be in scrollContainer.
 	const ScrollBody = rest.focusableScrollbar === 'byEnter' ? SpottableDiv : Fragment;
@@ -90,7 +94,12 @@ let Scroller = ({'aria-label': ariaLabel, hoverToScroll, ...rest}) => {
 			<ScrollContentWrapper {...scrollContainerProps} {...scrollContentWrapperRest}>
 				<ScrollBody {...focusableBodyProps}>
 					{rest.focusableScrollbar ? <ScrollbarPlaceHolder /> : null}
-					<UiScrollerBasic {...themeScrollContentProps} aria-label={ariaLabel} id={id} ref={scrollContentHandle} />
+					<UiScrollerBasic {...themeScrollContentProps} aria-label={ariaLabel} id={id} ref={scrollContentHandle}>
+						{(editable && direction === 'horizontal') ?
+							<EditableWrapper {...editableWrapperProps} /> :
+							children
+						}
+					</UiScrollerBasic>
 					{isVerticalScrollbarVisible ? <Scrollbar {...verticalScrollbarProps} /> : null}
 					{isHorizontalScrollbarVisible ? <Scrollbar {...horizontalScrollbarProps} /> : null}
 					{hoverToScroll ? <HoverToScroll {...hoverToScrollProps} /> : null}
@@ -201,6 +210,16 @@ Scroller.propTypes = /** @lends sandstone/Scroller.Scroller.prototype */ {
 	 * @public
 	 */
 	direction: PropTypes.oneOf(['both', 'horizontal', 'vertical']),
+
+	/**
+	 * Enables editing items in the scroller.
+	 * You can specify props for editable scroller as an object.
+	 * See the details in [EditableShape]{@link sandstone/Scroller.EditableShape}
+	 *
+	 * @type {sandstone/Scroller.EditableShape}
+	 * @public
+	 */
+	editable: EditableShape,
 
 	/**
 	 * Adds fade-out effect on the scroller.